move scales util code to scale service
